backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic & Settings
pydantic==2.7.1
pydantic-settings==2.3.0

# Cache / Broker
redis==5.0.4

# HTTP Client (async)
httpx==0.27.0

# Telegram Bot
aiogram==3.6.0

# Security
python-jose[cryptography]==3.3.0
pycryptodome==3.20.0

# Background Tasks
apscheduler==3.10.4

# Logging
loguru==0.7.2

# WebSockets
websockets==12.0

# Utils
python-dotenv==1.0.1
python-multipart==0.0.9
orjson==3.10.5

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0